Rename armv7 to arm32
