+++ src/trampolines/archdepend.h
@@ -1,5 +1,5 @@
 #ifdef AARCH64
 #include "aarch64/aarch64.h"
 #elif ARM
-#include "armv7/armv7.h"
+#include "armv32/armv32.h"
 #endif
